refactor pre executes
